Move devtools integration to a separate bundle
To enable using the React Developer Tools with preact without increasing
the size of the preact bundle, build the React DevTools integration
module as a separate bundle.
