+++ config/rollup.config.devtools.js
@@ -1,0 +1,15 @@
+import nodeResolve from 'rollup-plugin-node-resolve';
+import babel from 'rollup-plugin-babel';
+
+export default {
+	entry: 'devtools/index.js',
+	format: 'cjs',
+	plugins: [
+		babel({
+			sourceMap: true,
+			loose: 'all',
+			blacklist: ['es6.tailCall'],
+			exclude: 'node_modules/**'
+		})
+	]
+}
